update card popup flow
