--- card-pop-up-example/client/html/checkout.html
@@ -1,36 +1,0 @@
-<!doctype html>
-<html lang="en" xmlns="http://www.w3.org/1999/html">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Card Pop-up Example</title>
-    <link rel="stylesheet" href="./css/basePopup.css" />
-    <link rel="stylesheet" href="./css/sidebar.css" />
-    <link rel="stylesheet" href="./css/modal.css" />
-  </head>
-  <body>
-    <div class="wrapper">
-      <%- include("./partials/sidebar.html") %>
-      <main>
-        <h1>Checkout</h1>
-        <div class="checkout-item">
-          <img src="./assets/nice_glasses.jpeg" alt="Avatar" />
-          <h2>Rounded Glasses</h2>
-          <span>$0.57</span>
-        </div>
-        <div class="pay-btn-container">
-          <button id="pay-button" class="checkout">Pay</button>
-        </div>
-        <div id="notifications" role="alert"></div>
-        <div class="overlay" id="overlay">
-          <div class="popup">
-            <span class="close-btn" id="close-btn">✕</span>
-            <div id="popup-content"></div>
-            <button class="done-btn" id="done-btn">Done</button>
-          </div>
-        </div>
-      </main>
-    </div>
-    <script type="module" src="./checkout.js"></script>
-  </body>
-</html>
