+++ card-pop-up-example/client/html/css/basePopup.css
@@ -66,18 +66,14 @@ body {
   margin-left: 8px;
 }
 
-.order-status {
-  font-weight: 700;
-}
-
-.checkout-btn-container {
+.pay-button-container {
   display: flex;
   align-items: center;
   justify-content: center;
   margin-bottom: 10px;
 }
 
-.checkout {
+.pay-button {
   background-color: #4f55f1;
   border: none;
   border-radius: 12px;
@@ -88,41 +84,6 @@ body {
   display: inline-block;
   font-size: 16px;
   cursor: pointer;
-}
-
-.checkout-item {
-  display: flex;
-  align-items: center;
-  border: 1px solid #ddd;
-  padding: 16px;
-  border-radius: 8px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
-  width: 300px;
-}
-
-.checkout-item img {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  margin-right: 16px;
-}
-
-.checkout-item h2 {
-  flex: 1;
-  font-size: 18px;
-  font-weight: bold;
-}
-
-.checkout-item span {
-  color: #232625;
-  font-weight: bold;
-}
-
-.pay-btn-container {
-  display: flex;
-  align-items: center;
-  width: 300px;
-  padding: 16px;
 }
 
 #notifications {
