+++ card-pop-up-example/client/html/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Card Popup Example</title>
-    <link rel="stylesheet" href="css/basePopup.css" />
+    <link rel="stylesheet" href="./css/basePopup.css" />
     <link rel="stylesheet" href="./css/sidebar.css" />
     <link rel="stylesheet" href="./css/modal.css" />
   </head>
@@ -31,8 +31,8 @@
               <span class="price">$0.57</span>
               <span class="last-price">$1.19</span>
             </div>
-            <div class="checkout-btn-container">
-              <button id="checkout-btn" class="checkout">Checkout Now</button>
+            <div class="pay-button-container">
+              <button id="pay-button" class="pay-button">Pay</button>
             </div>
           </div>
         </div>
